+++ cx-portal-shared-components/src/components/index.tsx
@@ -85,6 +85,7 @@ export { Datepicker } from './basic/Datepicker'
 export { TransitionChip } from './basic/Chip/TransitionChip'
 export { Tooltips } from './basic/ToolTips'
 export { SortOption } from './basic/SortOption'
+export { ParentSubNavigation } from './basic/SubNavigation/ParentSubNavigation'
 
 export type { TableProps } from './basic/Table'
 export type {